后台计算价格

+++ service/src/main/java/cn/momia/service/promo/coupon/CouponService.java
@@ -12,6 +12,6 @@ public interface CouponService {
     int queryCountByUser(long userId, int status);
     List<UserCoupon> queryByUser(long userId, int status, int start, int count);
     boolean lockUserCoupon(long id, long orderId, long userCouponId);
-    UserCoupon getUserCoupon(int userCouponId);
+    UserCoupon getUserCoupon(long userCouponId);
     BigDecimal calcTotalFee(BigDecimal totalFee, Coupon coupon);
 }
